+++ SMH2-ejb/src/java/br/mack/fci/cc/smh/ejb/Evento.java
@@ -16,7 +16,7 @@ public class Evento {
     private Date _dataDoEvento;
     private String _sensor;
     private String _tipo;
-    private int _valor;
+    private double _valor;
     
     public Evento(){        
         _dataDoEvento = new Date(1900, 01, 01);
@@ -25,7 +25,7 @@ public class Evento {
         _valor = 0;
     }
        
-    public Evento(Date data, String sensor, String tipo, int valor) {
+    public Evento(Date data, String sensor, String tipo, double valor) {
         _dataDoEvento = data;
         _tipo = tipo;
         _sensor = sensor;
@@ -35,5 +35,5 @@ public class Evento {
     public Date getDataDoEvento() { return _dataDoEvento; }
     public String getSensor() { return _sensor; }
     public String getTipo() { return _tipo; }
-    public int getValor() { return _valor; }
+    public double getValor() { return _valor; }
 }
